paddleocr_service/pyworker/requirements.txt
# Vast.ai PyWorker SDK for serverless registration
vastai-sdk>=0.3.0
aiohttp>=3.10.0
psutil>=6.0
